+++ src/collective/documentgenerator/browser/generation_view.py
@@ -89,9 +89,12 @@ class DocumentGenerationView(BrowserView):
         rendered_document.close()
         os.remove(document_path)
 
-        # to avoid problems with long filename we take 120 first characters
-        first_part = u'{0} - {1}'.format(pod_template.title, safe_unicode(self.context.Title()))
+        # we limit filename to 120 characters
+        first_part = u'{0} {1}'.format(pod_template.title, safe_unicode(self.context.Title()))
         util = queryUtility(IFileNameNormalizer)
+        # remove '-' from first_part because it is handled by cropName that manages max_length
+        # and it behaves weirdly if it encounters '-'
+        first_part = first_part.replace('-', ' ')
         filename = '{0}.{1}'.format(util.normalize(first_part, max_length=120), output_format)
 
         return rendered, filename, gen_context