+++ src/collective/documentgenerator/tests/test_generation_view.py
@@ -155,14 +155,17 @@ class TestGenerationViewMethods(PODTemplateIntegrationTest):
         template_uid = pod_template.UID()
         generation_context = api.content.create(
             type='Folder',
-            title='Folder title long enough to test filename generation with accent héhé',
+            title='Folder title long enough to test filename generation 1.2.3.4 with accent héhé '
+                  'actually we test with a title that is longer than 120 characters so we may '
+                  'evaluate cropText method that is sensitive to some characters like -',
             id='test_folder',
             container=self.portal,
         )
         generation_view = generation_context.restrictedTraverse('@@persistent-document-generation')
         generation_view(template_uid, 'odt')
-        generated_id = 'general-template-folder-title-long-enough-to-test-filename-generation-with-accent-hehe'
-        msg = "File  {0} should have been created in folder.".format(generated_id)
+        generated_id = 'general-template-folder-title-long-enough-to-test-filename-generation-' \
+            '1-2-3-4-with-accent-hehe-actually-we-test-with-a-t'
+        msg = "File {0} should have been created in folder.".format(generated_id)
         self.assertTrue(generated_id in generation_context.objectIds(), msg)
 
         persistent_doc = generation_context.get(generated_id)
@@ -178,7 +181,8 @@ class TestGenerationViewMethods(PODTemplateIntegrationTest):
 
         self.assertEqual(
             filename,
-            u'General template - Folder title long enough to test filename generation with accent hehe.odt')
+            u'General template Folder title long enough to test filename '
+            u'generation 1.2.3.4 with accent hehe actually we test with a t.odt')
         self.assertEqual(content_type, 'application/vnd.oasis.opendocument.text')
 
     def test_persistent_document_generation_on_non_folderish_context(self):